+++ src/Gate.IO.Api/Spot/Responses/GateSpotPriceTriggeredOrder.cs
@@ -15,7 +15,7 @@ public class GateSpotPriceTriggeredOrder : GateSpotPriceTriggeredOrderRequest
     /// User ID
     /// </summary>
     [JsonProperty("user")]
-    public int UserId { get; set; }
+    public long UserId { get; set; }
 
     /// <summary>
     /// Creation time
